crypto: merge message modules
